+++ tests-adminata/App/Entity/Product.php
@@ -21,6 +21,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity]
 #[ORM\Table(name: 'demo_product')]
@@ -93,6 +94,7 @@ class Product implements \Stringable
 
     /** @var Collection<int, ProductVariant> */
     #[ORM\OneToMany(targetEntity: ProductVariant::class, mappedBy: 'product', cascade: ['persist'], orphanRemoval: true)]
+    #[Assert\Valid]
     private Collection $variants;
 
     public function __construct()